It is now possible to add and delete the instances of the FunctionControlPanel(aka the "Inputpackages"). The number of instances is still capped at 3 and has a minimum of 1 but this could be changed.
